requirements.txt
Django==1.9.7
django-daterange-filter==1.3.0
django-simple-captcha==0.5.1
gunicorn==19.6.0
meld3==1.0.2
MySQL-python==1.2.5
Pillow==3.2.0
six==1.10.0
supervisor==3.3.0
django-multiselectfield
